// src/postRepository/observers/enums.ts
export enum EnumPostActions {
  OnPostCreated = 'onPostCreated',
  OnPostUpdated = 'onPostUpdated',
  OnPostDeleted = 'onPostDeleted',
  OnPostFlagged = 'onPostFlagged',
  OnPostUnflagged = 'onPostUnflagged',
  OnPostReactionAdded = 'onPostReactionAdded',
  OnPostReactionRemoved = 'onPostReactionRemoved',
  OnPostApproved = 'onPostApproved',
  OnPostDeclined = 'onPostDeclined',
}